+++ distro/otel.go
@@ -28,6 +28,7 @@ import (
 	"os"
 	"strings"
 
+	"go.opentelemetry.io/contrib/instrumentation/runtime"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric/global"
@@ -199,6 +200,11 @@ func runMetrics(c *config, res *resource.Resource) (shutdownFunc, error) {
 
 	provider := metric.NewMeterProvider(o...)
 	global.SetMeterProvider(provider)
+
+	// Add runtime metrics instrumentation.
+	if err := runtime.Start(); err != nil {
+		return nil, err
+	}
 
 	return provider.Shutdown, nil
 }